Delay autosave and abort on an explicit cancel/save
Use ``setTimeout`` to delay the auto saving of a text assets.
Then cancel a pending auto save if the user clicks the save or
cancel buttons.  This avoids sending multiple requests, or having
to undo a save after clicking cancel.

Update the signature of ``SectionAssetEditView.saveModel()`` to
accept a delay (in milliseconds), and a success callback.  These
are both needed to implement this feature.

Addresses #848