Enable deletion of assets through REST API
This is needed to cover the case where a user creates a new text asset,
leaves the editor, triggering autosave, and later decides to click
the cancel button.

Without calling the ``destroy()`` method on the Backbone model and
deleting the model on the server-side as a result, the model will
still be associated with the story and show up in the asset drawer
when the story is reloaded.

Addresses #848