Destroy asset after disassociating with section
When a user clicks cancel, after a newly-added autosaved asset
has been created, it should be destroyed.  However, this needs
to happen in a callback after the server has successfully
disassociated the asset from the section, otherwise a race condition
occurs and an exception is thrown on the server.

The race condition occurs in
``StoryIndex.asset_relation_update_object``, which gets triggered
when a ``SectionAsset`` model is deleted. The race condition happens
if that method tries to access an asset that has already been deleted.

Update the signature of ``SectionView.removeAsset()`` to allow for
passing a ``success`` callback that will be executed after the
server responds. In ``SectionAssetEditView.cancel()`` use this
callback to inititate the request to remove the asset on the server.

Addresses #848