Fix editor size jumping after resizing
Fix a bug with the auto-expanding rich text editor where it was
shrinking after the user clicked outside the component element.
This was because of a couple of reasons.  First the toolbar was
toggling between shown and hidden states on blur/focus. Second, the
stylehost elements were getting set to the height of the content
and not the height + extra space.

Clean up the way default options and callbacks are handled in
``RichTextEditorMixin.createEditor()`` and add a ``toggleToolbar``
option that, if set to false, will always show the toolbar.

Addresses #844