Remove RichTextEditorMixin
This really didn't make sense as a mixin.  Instead, create a new
class, ``storybase.forms.RichTextEditor``, that extends
``wysihtml5.Editor`` but whose contstructor is like the old
``RichTextEditorMixin.createEditor()``.  This is also makes the
the editor more reusable across components, instead of just being
available in the builder.

Also clean up some of the internals for the initialization methods
of the editor.