Make sure body textarea is visible on edit
Make sure the ``body`` field's textarea element is visible when
editing a text asset before initializing the rich text editor.

This is needed because the wysihtml5 editor hides the textarea
by setting it's display property to ``none`` when it is initialized.
However, if the user clicks the "Cancel" button, the editor never
unhides the textarea element before the editor is destroyed.

The next time an editor is initialized, it copies the CSS properties
from the textarea, causing the editor to be hidden erroneously.