Trigger autosave on keyup and initialize form less often
Trigger text asset autosave on keyup events instead of the editor's
blur event.  This is needed because the editor wouldn't always lose
focus when we needed to autosave, such as when switching sections.
So, listen to keyup events on the editor element and pass them
through to the underlying textarea, which in turn triggers the
change event on the Backbone Form view, which we bind to with the
autosave logic.

In the process of implementing this, I found that we were calling
``SectionAssetEditView.initializeForm()`` all over the place.
This was problematic because one place where the form was initialized
was on the model's ``change`` event. Since we're autosaving the model,
the form would get clobbered each time the model autosaves. In general,
it's not great to have the overhead of recreating and re-rendering
the view when it's not needed.

As far as I can tell, it only needs to be called when we're ready
to edit the asset. So, I removed all the calls and added one to
``EditState.enter()``.

Addresses #848