Remove outdated failing JavaScript tests
As of 55408bd5c5f52ea963d6d8776f1a1a39d402f0b5, the form for
a ``SectionAssetEditView`` instance is initialized right before
rendering, so its no longer correct to test for it on initialization.

The behavior of widget.js is that empty storybase-story-list-embed
divs won't be processed as they don't have any URL providing elements
in them. This seems like fine behavior,
so remove the "replaces empty divs with a class of
'storybase-story-list-embed' with iframe ..." spec

jshint cleanup of spec files, especially lots of forgetting to include
'()'s in calls like ``new ClassName()`` and trailing semicolons.