Always show caption at bottom of asset and popup
Instead of moving the caption element to the top of the container,
copy its contents to a separate element and always show the caption
at the bottom.

This prevents content jumping around when hovering toggles the
popup. It also makes the caption immediately visible as the user
scrolls below the top of the image.

This approach no longer allows for a pure CSS solution as we need
to use JavaScript to create a separate element.

Cleanup trailing commas in object literal definition.

Addresses #717