Toggle editor display and save model in TitleView
In the Backbone view for editing the story title, toggle the
display of the title and the text input on blur and when the
Enter key is pressed and released.  Also set the title attribute
on the model to the new value of the input.

Automatically save the story model when its title attribute
changes.

Refactor the test code to be more DRY.

Addresses #578