Add a ``has-story`` class to the master view
When the user selects a template, ``AppView`` sets a ``has-story``
class on its element.

Right now, this is just used to toggle logo styles, but the
selecting template/editing story state seems like a significant
enough change that this class could be used to trigger other
style changes.

While ``ready:story`` seems like a more appropriate place to
set the class, we want to change the logo ASAP!

Addresses #578