Mock some models, libraries in JavaScript tests
``AppView`` does/triggers a bunch of stuff.  Now that the tests
are firing more events, we need to mock some libraries like
Modernizr and guiders to avoid errors.

Inside builder/views.js, check if the jQuery tooltipster plugin
exists and set it to a noop if it doesn't.  That way we can make
the plugin optional and not wrap a bunch of guards around calls
to the plugin.

Also remove the image toggling tests for ``TitleView``, in
anticipation of moving this functionality to a separate view.