Update the browser's location on initial save from editing title
When doing an initial edit of a story via the title or byline in
the header, be sure to make sure the browser's location updates
to contain the id of the story.

To do this, trigger a ``do:save:story`` event on the event bus instead of
calling ``this.model.save()`` if the story hasn't been saved
before.

The ``do:save:story`` event and the associated event handling seems
a bit dubious, but this is how we've been doing it.  Add a comment
that flags this pattern for review.

Addresses #578