Add close() method to BuilderView
Add a ``BuilderView.close()`` method to unbind model and event bus
events. This makes it easier to test views in an atomic way because
some of the models used in handling event bus events aren't always
mocked, especially after a particular spec is finished.  Basically,
this helps us clean up after ourselves in the ``afterEach()`` functions
in the specs.