Refactor next/previous navigation between workflow steps
Add ``WorkflowNextPrevView`` to handle navigation between workflow
steps instead of defining the next and previous navigation items
in the views that implement the different workflow steps.  Not
only should this be more DRY, it should also make it easier to
reorder workflow steps because this all happens in one place,
``AppView``.

Still need to wire this in to ``AppView`` and rip out the
initialization of individual instances of ``WorkflowNavView`` in
the views that implement the different workflow steps.

Addresses #579