Add "Story Information" workflow step
Add a "Story Information" workflow step that allows the user to
edit the story summary and featured image.  This replaces
the "Story Information" pseudo-section since the title and byline
can now be edited in the header.

The ``change()`` event handler was replicated in a number of views
and was always used alongside the ``saveAttr()`` method.  Move
the ``change()`` event handler into the ``StoryAttributeSavingMixin``
mixin.  Rename this mixin to ``ModelAttributeSavingMixin`` because
its also used for ``SectionEditView``.

Addresses #578, #579