Fix workflow step URLs when story is new
When a story is new, that is, it has no story_id because it has
never been saved, ``WorkflowStepView`` should generate links like
'/build/', '/build/info/' instead of '/build/undefined/info/'.

The logic for determining when the story ID should be included
was a little wonky.

Also make sure that the "Story Info" workflow step link doesn't
work if no model has been set, that is, if the template hasn't
been selected yet.

Addresses #579