Trigger initial save of story when featured asset is set
For an unsaved story, trigger the initial save of the story when
the featured asset is set.

Move the updating of the browser path out of ``BuilderView`` and into
``AppView``, allowing the path to reflect the current workflow step
and the app to stay on the "Story Info" workflow step if that's the
step the user was on when the save was initiated.

Make ``FeaturedAssetView`` re-delegate events when showing its
subviews.  This was breaking events in the sub-views.

As part of testing this, I mocked the license HTML endpoint as the
tests were previously making tons of requests to this endpoint for
some reason.

Addresses #579