Create a new character count view on re-render
When ``StoryInfoView`` is re-rendered, recreate the character counter
view.  Without doing this, view is bound to an editor instance that
no longer exists. It would be nice to be able to reuse this view, but
event binding/unbinding on the wysihtml5 editor requires us to track
all of our event handlers so we can unbind them later.  It seems
easier to just create a new view.

Refactor editor element setting and event binding of
``CharacterCountView``.  Pass the event bus object as an argument
to the constructor instead of accessing something global.

Addresses #579