Initialize workflow navigation views before BuilderView
Break fetching of story sections out of ``BuilderView.initialize()``
and into a separate ``initStory()`` method.  This allows
``BuilderView`` to be constructed and initialized before views
that need to listen to the ``ready:story`` event that is initiated
by ``initStory()``.

The motivation for this change is that ``WorkflowNextPrevView``
was not rendering with the story id in its links after the
initial save of a new story in the connected builder.  This was
because ``BuilderView.initialize()`` was causing ``ready:story``
to be triggered before ``WorkflowNavView`` was initialized.

Addresses #579