Show the story title input initially
For new stories, once a template is selected, show the title input
initially and its enhanced tooltip.

Also use the Tooltipster plugin to make the tooltip more noticeable.

In order to implement this, ``InlineEditorView.setModel()`` triggers
a ``set:model`` event on itself.

This change is meant to make the title input more prominent to avoid
confusion with the section title field.

Addresses #578