Remove styles for toggling the builder section list
Remove vestigal styles for controls that were used to show and hide
the section list (we've also called this the table of contents).

The section list has been always-visible for a very long time in the
app and there aren't any plans to make the section list hidable, so
remove the clutter of these unneeded styles.