Hide add new section icons when reordering sections
Use CSS to hide the spacers and add new section icons in the section
list when a section icon is being dragged to reorder the sections.

An extra "+" icon is still visible when the section icon is
"dropped", before the Backbone collection has finished syncing with
the server, triggering a re-render.  This is because the the CSS class
added by jQuery UI is removed before the Backbone collection syncs.
Fixing this is a more complicated fix that probably involves
optimistically updating the UI before the server syncs (and providing
a mechanism to undo).

Closes #905