Close old alerts when the publish button is pressed
Alerts triggered by the publish button should be cleared each
time the publish button is pressed.  Otherwise, the error for an
unsuccessful publish will remain, even after the missing fields
have been added and the story is successfully published.

In order to get this to work correctly, we need a mechanism for
views to track and close ``AlertView`` instances created as a
result of triggering an ``alert`` event on the event bus.  This
commit adds two changes that support this:

* ``AlertManagerView`` triggers a ``show:alert`` event on the event
  bus when it displays an alert.  The event parameters include
  the ``AlertView`` instance.
* The ``alert`` event now accepts an optional ``alertId`` property
  as part of the ``options`` argument.  This ID is bound to the
  created ``AlertView`` instance.

With these changes in place, views can provide alert IDs when
triggering ``alert`` and listen to ``show:alert`` and check the
view's ID to see if it is the view corresponding to the alert
triggered by the view.

Add an ``options`` argument to the ``initialize()`` method of
``PublishViewBase`` and its subclasses in anticipation of more
recent versions of Backbone that don't automatically extend
the view's ``options`` property with the options argument to
``initialize()``.

Addresses #579