Fix auto-setting of featured asset
Under some conditions, the asset model instance syncs with the
server before the event listener for the ``change:content`` is
wired up, causing the ``change:content`` event to not be fired,
and the featured asset to not be set.

In ``FeaturedAssetView.handleAddAsset()``, check if the ``content``
attribute is set, and if it is, immediately set the asset as the
featured asset instead of wiring up an event listener.

Closes #908