Scroll to top of window when switching sections or steps
In the builder, when the user changes workflow steps or story
sections, scroll the window to the top of the window.

Add ``window`` as a parameter wrapping the self-executing anonymous
function implementing the builder views.

I hoped that this would also work around the issue with drawer rendering
described in #904, but it doesn't and I'll need to address that
separately.

Addresses #910