Take scroll position into account for top margin
For the drawer and workflow step views, the ``margin-top`` CSS
property is set with JavaScript to accomodate the different
header heights between the "Build" and other steps.  This is
because of the header in the "Build" step also includes the section
list bar.

``AppView.pushDown()`` is called from ``AppView.render()`` in order
to push the subview and drawer container elements down below the
header.  To calculate the ``margin-top`` value, we need to take
into account the scroll position when calculating the offset of
the bottom of the header.

Addresses #904