Move title below workflow bar
A quick and dirty pass at UI changes intended to move the story
title closer to the story content and therefore make it easier
to identify.

* Remove the logo bar.
* Move the story title below the workflow step bar.
* Move the "tools" menu into a collapsible menu inside the workflow
  step bar.

There's still lots of cleanup work with this still to be done.

In particular, the event that triggers the class change that shows
the title bar seems to happen after the padding of the section/drawer
is calculated.

Addresses #579