Adjust subview and drawer padding when title is revealed
With the updated layout in 4fbf517ddea7a10afd817265b1dfa495b70a36ec,
the header height changes from the template selection step to the
other steps due to the story title bar being revealed.  In addition
to adding the ``has-story`` class (which will trigger CSS styles
that show the title bar) readjust the padding on the subview
container and drawer elements.

Addresses #579