Workflow step link colors
As per the request at
https://github.com/pitonfoundation/atlas/issues/579#issuecomment-30435243
update the workflow step link colors so they match site-wide navigation
and the active/available steps are more clear.

* Disabled steps are gray
* Available steps are white
* Current step is orange

The icon for the selected step stays white because I don't know the
location of the design file to create the hammer icon in orange.

Addresses #579