Force showing an updated version of the tour
Change the cookie that marks whether the user has seen the builder
tour, forcing users to see the tour again.  This is because the
changes to the workflow and the new placement of the title and
byline elements, and their editing behaviors, make it likely
that many users will need some reorientation.

Change the name of the cookie from ``storybase_show_builder_tour``
to ``storybase_builder_tour_seen`` and change its value from a
boolean value to a version string.  In the future, if the tour
needs to be updated and shown to users again, simply bump the
version number, stored in the ``TOUR_VERSION`` variable and
the tour will be shown again.

Addresses #579