Reassign assets to new containers on layout change
Instead of removing assets to the asset drawer when the layout of a
section changes, try to reassign them to corresponding containers in
the new layout.  When there is no corresponding container, e.g. when
there are fewer containers in the new layout, remove the asset to
the asset drawer as before.

In order to achieve this, I had to refactor some of the builder
Backbone code:

* Move the binding of sections to assets out of the view code and
  into the Asset model and the SectionAssets collection.
* Move the declaration/instantiation of the SectionAsset model
  out of the views.  This could likely be further improved by adding
  semantic names.
* Allow PUT requests to SectionAssetResource. This is needed to change
  the container when the layout changes.
* Define a mapping of container "weights" across different layouts.