Better API for managing Asset to Section relations
Manage these relationships through methods of the ``SectionAssets``
collection.  When a ``sync`` option is passed to
``SectionAssets.add()`` or ``SectionAssets.remove()``, persist
the relationship to the server through a ``SectionAsset`` model.

With this change, calling code no longer has to deal with the
``SectionAsset`` model instances and ``Asset`` model instances
don't need to know about their section relationship.  Calling code
just needs to do something that feels pretty natural.  Add/remove
models to/from a collection.  We provide on utility method on
the collection, ``updateSectionAsset`` for updating the container
when we change the layout or for persisting the relationship
after we've already added an item to the collection.

Still TODO:

* Better comments for the code

Addresses #914