Use a separate variable to toggle LESS in browser
For development, it makes sense to compile the LESS files in the
browser with less.js to make it easy to review changes.

I had been toggling this using ``settings.DEBUG``, but we also
need to be able to look at things in IE8, which isn't supported by
less.js.  Add a separate variable,
``settings.STORYBASE_USE_LESS_IN_BROWSER`` to toggle including the
LESS stylesheets (and less.js) in the template header.