Make "View Story" button appear higher up on detail page
Switch the order of the summary and featured image elements so
when they stack the summary (and view buttons) show up before
the image.  On wider displays, float the summary element right
to maintain the current look/feel.

Move the "View Story" button above the summary, so it doesn't get
hidden by long summaries.

Addresses #774, #921