Fix menu popups when menu items wrap to two lines
The fix is to adjust the padding of the <a> elements to position
the arrow and the padding of the ``.dd`` divs to position the boxes
when menu items wrap to the next menu.

This is a bit of a hack because I had to specify styles and breakpoints
for specific menu items.  This will break if menu items change length or
are added or reordered.  I thought about other solutions, like making
menu items fixed width on intermediate-width displays, but we get a lot
of rows due to the long menu titles.

I think the best workaround would be to add a little JavaScript
that detects when an item is wrapped and then assigns a CSS class
that fixes the positioning of the arrow/box.

I also made the padding for menu items a padding-right instead of
a padding-left so the first wrapped item will be aligned all the
way to the left.

I adjusted the styles for the "My Account" menu item so it only
floats right for wider displays.

Addresses #918