Dynamic toggling of sticky filters
Using a width-based media query to toggle sticky behavior doesn't
handle the case where the window is somewhat wide, but short.
This is exactly what happens when a mobile device is held in
landscape mode.

Instead, in the JavaScript scroll event handler, check to see if
the height of the filter container is less than a specified
percentage of the window height and only stick the filters if they
won't cover too much of the window.

Addresses #927