Responsive table styles.
Responsive table styles, using a similar strategy as Bootstrap's.

This especially targets the story list table on the "My Stories"
page.

To make a table responsive wrap it in a div with ``table-wrapper``
and ``responsive`` classes:

    <div class="table-wrapper responsive">
      <table>
        ...
      </table>
    </div>

I added some maximum value breakpoints and a mixin to implement this.

Since it's such a wide table, I had to make the breakpoint maximum
for the "My Stories" page wider than the default.

Addresses #774, #929