Return an error for too long dataset URLs
When a user tries to create a dataset by POSTing to
/api/0.1/datasets/assets/<asset_id>/, and the ``url`` field is
longer than 200 characters, respond with an error.

This is because the underlying model's URLField is implemented as
a varchar(200) in Postgres.  No automatic validation happens.

Also do some client-side validation as this is a better place to
give the user more clear feedback.

Addresses #940