Handle SSLError on oEmbed requests more gracefully
It appears that we were occasionally getting SSLError exceptions  when
hitting the Flickr oEmbed endpoint.  Handle these exceptions by
using the fallback behavior we already had defined for other types
of oEmbed failures.

Also cache oEmbed responses to help deal with this error.

Add the ``mock`` package as a development dependency and use it
in unit tests for this to patch ``ExternalAsset`` methods so it
raises ``SSLError``.

Addresses https://github.com/denverfoundation/storybase/issues/946